requirements.txt
boto3==1.9.217
botocore==1.12.217
colorama==0.4.1
Django==2.2.4
django-debug-toolbar==2.0
django-qr-code==1.0.0
django-widget-tweaks==1.4.5
docutils==0.15.2
jmespath==0.9.4
Pillow==6.1.0
psycopg2==2.8.3
PyQRCode==1.2.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
qrcode==6.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
PyVimeo==1.0.11
requests==2.22.0
django-settings-export==1.2.1
python-logstash==0.4.6
django-prometheus==1.1.0
xlwt==1.3.0
